+++ apps/frontend/src/pages/project/libs/components/project-details-menu/project-details-menu.tsx
@@ -1,15 +1,29 @@
-import { useNavigate } from "react-router-dom";
-
 import { Menu, MenuItem } from "~/libs/components/components.js";
-import { AppRoute } from "~/libs/enums/enums.js";
-import { configureString } from "~/libs/helpers/helpers.js";
-import { useCallback, usePopover } from "~/libs/hooks/hooks.js";
+import { AppRoute, PermissionKey } from "~/libs/enums/enums.js";
+import { checkHasPermission, configureString } from "~/libs/helpers/helpers.js";
+import { useAppSelector, useCallback, usePopover } from "~/libs/hooks/hooks.js";
 
 type Properties = {
+	onEdit: () => void;
 	projectId: number;
 };
 
-const ProjectDetailsMenu = ({ projectId }: Properties): JSX.Element => {
+const ProjectDetailsMenu = ({
+	onEdit,
+	projectId,
+}: Properties): JSX.Element | null => {
+	const { isOpened, onClose, onOpen } = usePopover();
+	const { authenticatedUser } = useAppSelector(({ auth }) => auth);
+
+	const handleEditClick = useCallback(() => {
+		onEdit();
+		onClose();
+	}, [onEdit, onClose]);
+
+	if (!authenticatedUser) {
+		return null;
+	}
+
 	const projectAccessManagementRoute = configureString(
 		AppRoute.PROJECT_ACCESS_MANAGEMENT,
 		{
@@ -17,20 +31,22 @@ const ProjectDetailsMenu = ({ projectId }: Properties): JSX.Element => {
 		},
 	);
 
-	const { isOpened, onClose, onOpen } = usePopover();
-	const navigate = useNavigate();
-
-	const handleManagementClick = useCallback(() => {
-		navigate(projectAccessManagementRoute);
-	}, [navigate, projectAccessManagementRoute]);
+	const hasManageAccessPermission = checkHasPermission(
+		[PermissionKey.MANAGE_ALL_PROJECTS],
+		authenticatedUser.groups.flatMap((group) => group.permissions),
+	);
 
 	return (
 		<Menu isOpened={isOpened} onClose={onClose} onOpen={onOpen}>
-			<MenuItem
-				iconName="access"
-				label="Manage Access"
-				onClick={handleManagementClick}
-			/>
+			<MenuItem iconName="pencil" label="Edit" onClick={handleEditClick} />
+
+			{hasManageAccessPermission && (
+				<MenuItem
+					href={projectAccessManagementRoute}
+					iconName="access"
+					label="Manage Access"
+				/>
+			)}
 		</Menu>
 	);
 };