+++ apps/frontend/src/pages/project/project.tsx
@@ -1,18 +1,24 @@
 import {
 	Breadcrumbs,
 	Button,
+	Modal,
 	PageLayout,
 } from "~/libs/components/components.js";
 import { AppRoute, DataStatus } from "~/libs/enums/enums.js";
 import {
 	useAppDispatch,
 	useAppSelector,
+	useCallback,
 	useEffect,
 	useModal,
 	useParams,
 } from "~/libs/hooks/hooks.js";
-import { actions as projectActions } from "~/modules/projects/projects.js";
+import {
+	actions as projectActions,
+	type ProjectPatchRequestDto,
+} from "~/modules/projects/projects.js";
 import { NotFound } from "~/pages/not-found/not-found.jsx";
+import { ProjectUpdateForm } from "~/pages/projects/libs/components/components.js";
 
 import {
 	ContributorsList,
@@ -29,6 +35,7 @@ const Project = (): JSX.Element => {
 		project,
 		projectContributors,
 		projectContributorsStatus,
+		projectPatchStatus,
 		projectStatus,
 	} = useAppSelector(({ projects }) => projects);
 	const {
@@ -37,12 +44,41 @@ const Project = (): JSX.Element => {
 		onOpen: onSetupAnalyticsModalOpen,
 	} = useModal();
 
+	const {
+		isOpened: isEditModalOpen,
+		onClose: handleEditModalClose,
+		onOpen: handleEditModalOpen,
+	} = useModal();
+
 	useEffect(() => {
 		if (projectId) {
 			void dispatch(projectActions.getById({ id: projectId }));
 			void dispatch(projectActions.loadAllContributorsByProjectId(projectId));
 		}
 	}, [dispatch, projectId]);
+
+	useEffect(() => {
+		if (projectPatchStatus === DataStatus.FULFILLED) {
+			handleEditModalClose();
+
+			if (projectId) {
+				void dispatch(projectActions.getById({ id: projectId }));
+			}
+		}
+	}, [projectPatchStatus, handleEditModalClose, dispatch, projectId]);
+
+	const handleEditProject = useCallback(() => {
+		handleEditModalOpen();
+	}, [handleEditModalOpen]);
+
+	const handleProjectEditSubmit = useCallback(
+		(payload: ProjectPatchRequestDto) => {
+			if (project) {
+				void dispatch(projectActions.patch({ id: project.id, payload }));
+			}
+		},
+		[dispatch, project],
+	);
 
 	const isLoading =
 		projectStatus === DataStatus.PENDING || projectStatus === DataStatus.IDLE;
@@ -75,7 +111,11 @@ const Project = (): JSX.Element => {
 					<div className={styles["project-layout"]}>
 						<div className={styles["project-header"]}>
 							<h1 className={styles["title"]}>{project.name}</h1>
-							<ProjectDetailsMenu projectId={project.id} />
+
+							<ProjectDetailsMenu
+								onEdit={handleEditProject}
+								projectId={project.id}
+							/>
 						</div>
 
 						<div className={styles["project-description-layout"]}>
@@ -101,6 +141,17 @@ const Project = (): JSX.Element => {
 							/>
 						</div>
 					</div>
+
+					<Modal
+						isOpened={isEditModalOpen}
+						onClose={handleEditModalClose}
+						title="Update project"
+					>
+						<ProjectUpdateForm
+							onSubmit={handleProjectEditSubmit}
+							project={project}
+						/>
+					</Modal>
 
 					<SetupAnalyticsModal
 						isOpened={isSetupAnalyticsModalOpened}