+++ scripts/analytics/src/index.ts
@@ -1,5 +1,5 @@
+#!/usr/bin/env node
+
 import { analyticsCli } from "./libs/modules/analytics-cli/analytics-cli.js";
 
 analyticsCli.init();
-
-export { getPackageName } from "./libs/modules/analytics-cli/analytics-cli.js";