+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.22.1](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/shared-v1.22.0...shared-v1.22.1) (2024-09-18)
+
+
+### Bug Fixes
+
+* fix check has permission helper gf-248 ([#378](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/378)) ([ffc88da](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/ffc88da2ec88eb24b7e6e0dabe5f0b4ddd96a3ab))
+
+## [1.22.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/shared-v1.21.1...shared-v1.22.0) (2024-09-17)
+
+
+### Features
+
+* add permission to manage all projects gf-248 ([#282](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/282)) ([0bc48e9](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/0bc48e980a5aacf3b43cab5aacf2c9583c36a3c2))
+
 ## [1.21.1](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/shared-v1.21.0...shared-v1.21.1) (2024-09-17)